backend/requirements.txt
# FastAPI Stack
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# LangChain stack
langchain==1.2.0
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0

# LangGraph
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-sdk==0.3.1

# OpenAI
openai==2.15.0
tiktoken==0.12.0

# Tavily
tavily-python==0.7.17

# Database (psycopg3)
SQLAlchemy==2.0.45
psycopg[binary]==3.3.2

# HTTP
httpx==0.28.1

# SSE
sse-starlette==2.1.3